refactor: BoardロケーションDTOの一元化
