+++ src/features/board/application/dto/BoardLocationDTO.ts
@@ -1,0 +1,15 @@
+import type {
+  BoardStatus,
+  TrustLevel,
+} from "@/shared/domain/board/BoardAttributes";
+
+export interface BoardLocationDTO {
+  id: string;
+  boardNumber: number | null;
+  name: string | null;
+  address: string;
+  longitude: number | null;
+  latitude: number | null;
+  status: BoardStatus;
+  trustLevel: TrustLevel;
+}
